Moved some function to the main python query and added cleanup
